+++ packages/patternfly-4/react-table/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.13.51](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@2.13.50...@patternfly/react-table@2.13.51) (2019-07-05)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
+## [2.13.50](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@2.13.49...@patternfly/react-table@2.13.50) (2019-07-05)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+
+
+
+
 ## [2.13.49](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@2.13.48...@patternfly/react-table@2.13.49) (2019-07-05)
 
 **Note:** Version bump only for package @patternfly/react-table